Migrate code to a main component - Best practice
